+++ CHANGES.markdown
@@ -1,3 +1,6 @@
+## Changes in next
+ - `Show` instance for `Fingerprint`
+
 ## Changes in 0.2.0
  - Drop GHC 6.12 (and `base-4.2.0.0`) compatibility
  - Fix Windows, GHCJS build